split unit tests into hmf and lss
